Refactor gatsby-node.js in gatsby-theme-blog GH-354
