+++ .changeset/yellow-walls-reflect.md
@@ -1,0 +1,5 @@
+---
+'@undataforum/gatsby-theme-blog': patch
+---
+
+Refactor gatsby-node.js
