--- themes/gatsby-theme-blog/gatsby/on-create-node.js
@@ -1,135 +1,0 @@
-const path = require('path');
-
-const { createFilePath } = require('gatsby-source-filesystem');
-const remark = require('remark');
-const strip = require('strip-markdown');
-
-const withDefaults = require('../utils/default-options');
-const urlResolve = require('../utils/url-resolve');
-
-module.exports = (
-  { node, actions, getNode, createNodeId, createContentDigest },
-  themeOptions
-) => {
-  const { basePath, collection } = withDefaults(themeOptions);
-  // Process MDX nodes only.
-  if (node.internal.type !== `Mdx`) {
-    return;
-  }
-
-  // Parent fileNode makes `name` option from `gatsby-source-filename` available as `sourceInstanceName`.
-  const fileNode = getNode(node.parent);
-  const name = fileNode.sourceInstanceName;
-
-  // Process files in `contentPath` location only.
-  if (name === collection) {
-    const { createNode } = actions;
-
-    // Process description.
-    let description;
-    if (node.frontmatter.description) {
-      description = node.frontmatter.description;
-    } else {
-      // Four scenarios for which we need to match first paragraph:
-      // - multiple paras with import statement
-      // - multiple paras without import statement
-      // - one para with import statement
-      // - one para without import statement
-      // Regex:
-      // - match subsequent non-empty lines (but not lines starting with "import")
-      // - lookbehind and there should be two line feeds (\n)
-      const match = node.rawBody.match(/(?<=\n{2})((?!import).+\n)+/);
-      if (match) {
-        description = match[0];
-      }
-    }
-
-    // Use this ID to link node that processes any Markdown in description.
-    const descriptionNodeId = createNodeId(
-      `${collection}-description-${description}`
-    );
-
-    // Use this ID to link node that processes Markdown in title.
-    const titleNodeId = createNodeId(
-      `${collection}-title-${node.frontmatter.title}`
-    );
-
-    // Process path and slug.
-    let slug;
-    if (node.frontmatter.slug) {
-      slug = node.frontmatter.slug;
-    } else {
-      // relativePath in corresponding file node is relative to contentPath from corresponding gatsby-source-filesystem config.
-      // Therefore, pass in '' for basePath (argument basePath is different from theme option basePath.
-      slug = path.basename(createFilePath({ node, getNode, basePath: '' }));
-    }
-    const href = urlResolve(basePath, slug);
-    const post = {
-      slug,
-      collection,
-      // Foreign key reference to node that will be created further down.
-      title___NODE: titleNodeId,
-      date: node.frontmatter.date,
-      // Contains author slugs.
-      authors: node.frontmatter.authors,
-      images: node.frontmatter.images,
-      // Foreign key reference to node that will be created further down.
-      description___NODE: descriptionNodeId,
-      path: href,
-    };
-    const postNode = {
-      ...post,
-      // Generated ID is namespaced to plugin.name.
-      id: createNodeId(`${collection}-${slug}`),
-      // Make post node aware of MDX node.
-      parent: node.id,
-      children: [],
-      internal: {
-        type: 'Post',
-        contentDigest: createContentDigest(post),
-      },
-    };
-    createNode(postNode);
-
-    // Create description node that processes Markdown in description.
-    // https://www.christopherbiscardi.com/post/creating-mdx-nodes-from-raw-strings/
-    const descriptionNode = {
-      id: descriptionNodeId,
-      parent: postNode.id,
-      children: [],
-      internal: {
-        type: `${collection}Description`,
-        contentDigest: createContentDigest(description),
-        mediaType: 'text/markdown',
-        content: description,
-      },
-      // Strip Markdown.
-      text: remark()
-        .use(strip)
-        .processSync(description)
-        .contents.replace(/\n/g, ' ')
-        .trim(),
-    };
-    createNode(descriptionNode);
-
-    // Create title node that processes Markdown in title.
-    // https://www.christopherbiscardi.com/post/creating-mdx-nodes-from-raw-strings/
-    const titleNode = {
-      id: titleNodeId,
-      parent: postNode.id,
-      children: [],
-      internal: {
-        type: `${collection}Title`,
-        contentDigest: createContentDigest(node.frontmatter.title),
-        mediaType: 'text/markdown',
-        content: node.frontmatter.title,
-      },
-      // Strip Markdown.
-      text: remark()
-        .use(strip)
-        .processSync(node.frontmatter.title)
-        .contents.trim(),
-    };
-    createNode(titleNode);
-  }
-};
